requirements.txt
flask
flask-sqlalchemy
flask-session
python-dotenv
bcrypt
